Summary: Added initial prompt for NEW or SAVED plan; also changed time entry from ET to UTC.
PlanningLViewFactory.java:
1. Modified the logic in the createLView() method to first prompt for New Plan or Load Saved Plan.
2. Added method displayNewOrLoadDialog().
3. Modified method displayPlanningLayer Dialog() to accept start and end times in UTC, and then converts them to ET for internal use.

PlanningLView.java:
1. Added method viewCleanup() that gets called when the planning layer is deleted. It's only functionality at this time is to remove all loaded SPICE kernels.

SavedPlan.java (NEW)
This class will be used to save and restore a plan. The actual save and restore logic has yet to be implemented.

PlannedSurfaceFeature.java (NEW)
This class extends SurfaceFeature, adding a specific time stamp for the surface feature. It will be used when constructing plans (for both planned imaging targets or planned ground station fly-overs), and represents some of the contents of a SavedPlan object.